main/postmarketos-demos, -ui-weston: disable x86
Weston does not build for x86, because it depends on libunwind. And
libunwind is not available in Alpine for x86. I've also tried to
simply enable x86 in libunwind, but it does not build out of the box.

In order to start building again for x86 at all, let's disable weston
and all packages that depend on it on for now.