postmarketos-ui-weston: remove --pixman-type hack
Do not add --pixman-type to the commandline, when
deviceinfo_weston_pixman_type is filled out.

--pixman-type was enabled in Weston with a custom patch, that currently
prevents us from upgrading Weston (#136).

The option allowed working around broken framebuffer drivers in
Android downstream kernels, which reported the wrong color format.
But it only works for Weston, the right way to patch this would be
patching the kernels, and we have some approaches here:
https://wiki.postmarketos.org/wiki/Troubleshooting:display#My_screen_is_red.21