Revert "linux-*: use upstream gcc6 [skip ci]"
This reverts commit ee659a5bb4ad94f33ec47e6ae7bf9e856f627155
and increases the pkgrels of all affected linux pmaports.

I have compiled *every single kernel* that was modified with this
commit, and it worked. That took 12 hours. So I'm pretty confident that
this is a good commit. Let's roll it out and go back to stability \o/

I'll kick off the binary repo building directly after pushing this, but
it will take some time until all binary packages are available again.

[skip ci]: it wouldn't finish in time.