msm-fb-refresher: also do ioctl on errors (!89)
I noticed that some times the framebuffer driver gets configured in a
way that the ioctl performed by msm-fb-refresher returns something lower
than zero:

https://github.com/LineageOS/android_kernel_samsung_msm8974/blob/0ed263db093b1fc590589836ee36956c327998c3/drivers/video/fbmem.c#L877-L911

For example when the Xorg starts I noticed it does a ioctl BLANK and
UNBLANK, but if msm-fb-refresher performs a ioctl in that moment it
stops its loop and exit.

For this reason I lost a lot of time trying to understand what was the
problem with Xorg not displaying anything until I noticed that I had
to restart the msm-fb-refresher.

With this change we don't have to care about msm-fb-refresher as it will
continue call ioctl PAN even if the framebuffer returns some negative
error code.