kde/itinerary, kde/kitineary: remove (!156)
Right now, this doesn't build anymore, it is failing with:

/home/pmos/build/src/itinerary-c47772e49de6662acd7769d5c8cef94379ea7bcf/src/app/reservationmanager.h:21:10: fatal error: KItinerary/ExtractorRepository: No such file or directory
 #include <KItinerary/ExtractorRepository>
          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

I've tried to update it to the latest commit, but then it fails with
another error. Turns out, this package depends on kitinerary (with a
k!), and whenever we update that, we also need to update this package
to the latest commit. Unfortunatelly, itinerary does not have tagged
versions yet.

So I would like to remove it for now, it does not seem like anybody is
using it at this point, and so it's just additional maintenance effort
for nothing. We could package it again, when it has tagged versions and
is therefore easy to update together with kitinerary.

Remove kitinerary too, because it isn't used by anything else.