google-crosshatch: add WIP mainline kernel (!226)
This is tested to boot on google-crosshatch, following these
instructions:

$ pmbootstrap init, choose the mainline kernel
$ pmbootstrap install --no-fde
$ pmbootstrap initfs hook_add debug-shell
(enter fastboot mode on the device)
$ fastboot set_active a
$ fastboot flash dtbo_a ~/.local/var/pmbootstrap/chroot_rootfs_google-crosshatch/boot/dtbo-google-crosshatch-mainline.img
$ pmbootstrap flasher boot

And you should get telnet after a while. This kernel takes a long time
to build, so [skip ci]. I can confirm it builds locally.

Note that this patch adds the mainline dtb path to deviceinfo. But the
parameter gets ignored on the downstream kernel, so it's fine to keep it
in the shared deviceinfo.

Mainlining this device is work in progress, details at:
https://gitlab.com/postmarketOS/pmaports/issues/153