linux-purism-librem5dev: Disable CONFIG_FW_LOADER_USER_HELPER (!237)
When working on Librem5 devkit we had issue with touchscreen driver not
probing fast enough and taking ~70 seconds to finish initialization.

Touchscreen driver for the Librem5 devkit uses goodix driver, roughly
speaking device initialization is done in following sequence,

- start kernel and load driver
- attempt to request firmware
- if no firmware is found use defaults value
- initialize touchscreen

Now, on Librem 5 no firmware is provided (of course!) which makes 2nd
step 4 step process,

- attempt to load firmware at initramfs level
- fail, wait for it userspace to finish loading
- userspace attempts to load firmware
- fail and tell kernel to stop looking for firmware.

Step 2 to 4 is done after udev is loaded in userspace, which easily
takes 40-50 seconds. Since Purism Librem 5 devkit is never going to have
userspace level firmware, it is safe bet to disable this altogether.

[skip ci] tested locally and this is going to take more then hour on CI
and fail