main/wireguard-postmarketos-stable: remove
This isn't building anymore, because linux-postmarketos-stable has been
switched from armhf to armv7 in pma!240.

We could change the arch of wireguard-postmarketos-stable too, but the
package is not maintained at all, and I doubt that anybody is using it.
So I'm dropping it. If there are any users, please step up to maintain
the package.