ouya: blacklist partitions to avoid brick (!247)
Thanks to Robert Yang's pmbootstrap patch, we can now avoid flashing
anything to the boot and recovery partition of the Ouya. With this
patch, the latest pmbootstrap will refuse to do so with:

ERROR: 'boot' partition is blacklisted from being flashed! See the Ouya device wiki page for more information.