asus-grouper: modernize device package (!253)
Updated the device-asus-grouper package to use devicepkg-dev.

I've tested the resulting image and it builds, boots and has a working
touchscreen (in Weston)

I've also noticed that the device reboots itself at times, both in the
USB ssh console and in Weston. This seems to be a known issue but I
don't know how prevalent it was before.