main/postmarketos-ui-weston: fix custom configs (!423)
Use the start_weston.sh wrapper script again, and make it work with
lightdm. It makes custom weston configs possible, and allows starting
postmarketos-demos together with weston. The custom config logic uses a
default config from postmarketos-ui-weston unless it is overwritten
with a device specific config. Our default configs use fbdev in most
cases, as this is not the case since the change to lightdm, weston was
broken for most devices with downstream kernels, that only work with
framebuffer.

Lastly, set myself as maintainer for this UI. Resolves #289

Ollieparanoid extended this patch and made sure that this boots up
weston on the samsung-i9100 again.