temp/mesa: fix install prefix, enable release optimizations (!431)
The description below is a quote from Brian Masney in #296.

The pmOS mesa packages need to have -Db_ndebug=true passed to the meson
build so that debug assertions and extra validations are disabled to
improve overall system performance. See [1] for more details. Fedora
also recently made this change: [2]

This change, along with using the master branch of mesa, gets the
Freedreno driver working properly with X11 on the Nexus 5 and presumably
other A3XX-based systems.

--buildtype=release should also be passed to meson for extra
optimizations. See the thread [3] for more details.

[1] https://gitlab.freedesktop.org/mesa/mesa/blob/master/docs/meson.html#L321
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1692426.
[3] https://lists.freedesktop.org/archives/mesa-dev/2019-June/219960.html

[ci:skip-build]: already built successfully in CI