phosh/squeekboard: new aport (!523)
Squeekboard is going to replace virtboard in Phosh. However, it requires
Rust which hasn't yet been packaged on Alpine Linux for aarch64. So for
now, keep virtboard as the default (it still works fine) and update it
to the latest commit (see follow-up patch), but also introduce
squeekboard so we can test it and eventually enabled it for aarch64 and
replace virtboard.