temp/{networkmanager,wpa_supplicant}: new aports (!547)
Fork from Alpine to apply Martijn's patch:
https://github.com/alpinelinux/aports/pull/9894

"This makes wpa_supplicant run in dbus mode if no config file has been
created for it, fixing the networkmanager integration. This config has
been used for a long time in postmarketOS and works fine.

It also removes the iwd dependency from networkmanager again so
wpa_supplicant is used because iwd doesn't support as many chipsets as
wpa_supplicant and it doesn't run on older kernels."

[ci:skip-build]: already built successfully in CI for x86_64, I've
                 tested that it builds for armhf and aarch64 too.