phosh/squeekboard: disable aarch64, x86
Both architectures don't have a rust package in Alpine, therefore
squeekboard cannot be built. I did not notice this before, because the
phosh dir was not explicitly listed for building packages in the binary
repository. Instead, only postmarketos-ui-phosh was built with all
depends for all arches, but this package does not depend on squeekboard.

[ci:skip-vercheck]: only arch was changed