oppo-find-7a: linux: fix depends in dev subpackage
Insert missing space into the depends line, so it reads "gmp-dev bash"
instead of "gmp-devbash". This did not cause a build failure, but made
the upstream compat CI check fail after the binary package had been
built and uploaded. It would also cause an error if one tried to install
the package. The CI check will pass once the fixed package is built and
uploaded.

[ci:skip-build]