temp/qt5-qtwayland: remove non-applying patches
I did not check if patches still applied yesterday, and it turns out
that almost all of them don't apply anymore (probably because the
changes are in the upstream source now?).

Note that I did not test if plasma mobile is working correctly again
with this version, I'm just making the build errors go away (which I
did not notice right away yesterday, as qt5-qtbase was still building
and we had to upgrade that anyway, so there was no point in waiting).

Build tested and working for x86_64.

Fixes: 86a0ecc04a8dd41f33e937a9b9a53c92f8ff8dfb ("temp/qt5-qt*: upgrade to 5.12.5")