bpo/submit.py: require and pass BPO_JOB_{ID,NAME}
Let's send both to to the bpo server with each request. This is needed
to handle failing jobs (we need to know which job exactly is failing).

Note that this is not directly using a sourcehut specific environment
variable, because with bpo it is also possible to choose a "local"
job service for testing (manual and testsuite). The local job service
will just generate a random number as job id and place that in this
variable.