main/postmarketos-mkinitfs: configure USB gadget serialnumber (!684)
Currently, the USB networking setup using configfs does not configure
a "serialnumber" for the USB gadget. This means that an empty
serial number will be sent during USB enumeration.

This is usually not a problem, but it seems to cause issues under
special circumstances (on asus-me176c on one of my PCs).

The USB Gadget configfs documentation also suggests setting
the "serialnumber":
  https://www.kernel.org/doc/Documentation/usb/gadget_configfs.txt
It can be any string, so we can just set it as "postmarketOS" like
the manufacturer.