CI: skip git versions in framework check
The framework check is supposed to make sure, that we always have the
same versions of multiple packages that belong to the same framework.
We must exclude git versions for this check to have it working properly.

The interesting question is of course: why was this working before?

The check is using the url field to figure out which packages belong to
a certain framework, and by chance, all packages that had a git version,
did also use single quotes around the URL:
	url='http://qt-project.org/'
With the old APKINDEX parsing code, this was parsed as:
	url: "'http://qt-project.org/'"
And therefore, this was considered a different URL than the URL with
double quotes (again, by chance!) used by the stable qt versions:
	url: "http://qt-project.org/"
I have improved the APKBUILD parsing code in pmbootstrap today, and now
it is properly parsing both URLs.

Related: https://gitlab.com/postmarketOS/pmbootstrap/merge_requests/1837