mrhlpr.json: fix matching of numbers in regex (!957)
In the regex [a-z-0-9] you don't need a backslash before the minus in
the middle, so remove that to bring it in line with other regexes.

In the regex [a-z\/-0-9] you do need a backslash in front of the minus
though, otherwise only the numbers 0 and 9 are allowed by the regex.