google-crosshatch: fix build with downstream kernel (!1016)
The downstream kernel doesn't provide a separate dtb file,
but the experimental mainline kernel does.

Previously, the deviceinfo specified a dtb file for both kernels,
and pmbootstrap ignored the missing file for downstream kernel.

pmbootstrap now errors when the dtb file is missing, so this
fixes the downstream kernel by specifying a dtb file only for
the upstream kernel.