qemu-*: drop linux-postmarketos-mainline/stable kernel option (!1029)
We haven't been maintaining linux-postmarketos-mainline/stable very
well. Let's recommend use of one of the Alpine kernels which are actively
updated with new LTS updates (see #464).

This simplifies upcoming refactoring of the QEMU device ports,
which may (or may not) work with the more minimal kernel configurations
in linux-postmarketos-mainline/stable.