qemu-*: cleanup, use consistent hardware (!1029)
pmbootstrap now configures consistent QEMU hardware for all architectures.
This means that the device package for amd64/aarch64 is now almost identical,
except for the architecture and the name of the serial port.

pmbootstrap always uses virtio-gpu now (but not necessarily virgl),
so there is only one consistent set of initfs modules we need to configure
for both architectures.

Also add the xfce4 package from amd64 to aarch64, exactly the same
problem with the wallpaper exists there.