device: move .shared-patches to device/.shared-patches (!1068)
All the downstream kernels reference ../../.shared-patches in their
symlinks. Now that we have moved device ports to device/testing/*,
those symlinks are no longer working.

Changing the path would require fixing all downstream packages *and*
a new pmbootstrap release. Overall it seems easier to move the
.shared-patches folder to device/.shared-patches.

Since it is only relevant for downstream kernel packages in device/
that might be a better place in general.