main/umockdev: remove
Left over from unity 8, which has been removed from pmaports earlier.
umockdev depends on libgudev in Alpine, which has been moved from main
to community. As of writing, there is no armhf package, so the upstream
compatibility test is failing. There is no point in waiting until the
package was built though, since nobody is currently using in in
pmaports.