wileyfox-crackling: remove downstream kernel (MR 1125)
The device port gets increasingly hard to maintain with all the differences
between the downstream and the mainline kernel. Especially deviceinfo
has quite some duplication only because the downstream kernel needs to be
handled as well.

The mainline port has much more features working. The only limitation
compared to the downstream kernel is the lack of a proper charging driver.
But even on downstream postmarketOS the battery driver is not working
properly for some reason.

For charging, a downstream installation in recovery seems more suitable
at the moment. Eventually we will build some functionality into the
bootloader to load a simple downstream initramfs when the device gets
booted for charging.