main/linux-postmarketos-allwinner: enable CONFIG_CGROUP_FREEZER (MR 1569)
Enable the freezer cgroup. This allows a user to start a large application like
firefox, freeze it to save some power, and unfreeze it to continue using it,
without having to restart the application completely. Stopping and restarting
the application completely results in more battery usage as well as lost state.

Signed-off-by: Serge Hallyn <serge@hallyn.com>