xiaomi-wt88047: add mainline port for Xiaomi Redmi 2 (MR 1604)
Replace the old downstream port with a new aarch64 port running the
linux-postmarketos-qcom-msm8916 close-to-mainline(TM) kernel.
Most of the functionality is working:

  - USB Network
  - Storage (eMMC, SD card)
  - Display, brightness control
  - Touchscreen, buttons
  - IMU (accelerometer+gyroscope)
  - Notification LED
  - Sound (Speaker, Headphones, Microphones)
  - WiFi, Bluetooth
  - Modem (SMS, voice calls, mobile data)

Coming soon (hopefully): Battery/Charging, USB-OTG (not working yet)

There are many variants of the device, so more changes to handle
differences between them may be needed in the future. For now,
the only relevant difference is the required modem firmware:

  - wt88047 (global variant)
  - wt86047 (China variant)

... have different kind of modems and they do not seem to work with
the firmware of the other one. The way this is handled in this device
package is that there is a separate kernel variant for both of them.
All other required firmware can be shared, and therefore there is no
differentiation needed when using the kernel variant without modem.

I renamed "wingtech-wt88047" to "xiaomi-wt88047".
My reasoning for that is the following:

  - Wingtech is the ODM (original design manufacturer), they designed the
    hardware and (probably) manufactured it. But postmarketOS port applies
    to the combination of both hardware and firmware, and the firmware was
    at least partially provided by Xiaomi.

  - wt88047 looks lonely with the "wingtech" vendor, when all other Xiaomi
    devices use "xiaomi", even though many from them were probably also
    designed by Wingtech.

  - We don't use the ODM hardware name for other devices either, e.g.
    "wileyfox-crackling" would be "longcheer-l8150" with the same approach.

I think the port is better visible as "xiaomi-wt88047" (too bad Xiaomi
seemingly did not have a proper codename for this device...)