temp/wlroots: revert 0 dimension error check (MR 1702)
Currently phosh is broken in postmarketOS edge. Fork wlroots from Alpine
and revert the new consistency check that results in the breakage. I've
submitted the same to Alpine, but let's get it in pmOS now so it is
fixed ASAP.

Alpine MR: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/14522
[ci:skip-vercheck]: new package added with -r1 on purpose